requirements/base.freeze.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
hiredis==2.3.2
idna==3.7
kombu==5.3.7
msgpack==1.0.8
pillow==10.3.0
promise==2.3
prompt-toolkit==3.0.43
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
redis==5.0.4
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
